Allow to add multiple videos for one decorated url
